Revert "iTorrent framework replaced"
This reverts commit 9b05d78431e1027693a3970668f2badf560c0753.

# Conflicts:
#	iTorrent.xcodeproj/project.pbxproj
